Kernel: Remove problematic memory mapping methods
mmap() & mmap_region() methods are removed from ACPI & DMI components,
and we replace them with the new MM.allocate_kernel_region() helper.

Instead of doing a raw calculation for each VM address, from now on we
can use helper functions to do perform those calculations in a neat,
reusable and readable way.

+++ Kernel/PCI/Access.cpp
@@ -135,7 +135,7 @@ void disable_bus_mastering(Address address)
 {
     PCI::Access::the().disable_bus_mastering(address);
 }
-u32 get_BAR_Space_Size(Address address, u8 bar_number)
+size_t get_BAR_Space_Size(Address address, u8 bar_number)
 {
     return PCI::Access::the().get_BAR_Space_Size(address, bar_number);
 }
